Delete Models folder and move it to Models class
